refactor(frontend): Move ButtonHero to Hero folder
