+++ src/frontend/src/btc/components/convert/ConvertToCkBTC.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { BTC_MAINNET_TOKEN } from '$env/tokens.btc.env';
 	import IconCkConvert from '$lib/components/icons/IconCkConvert.svelte';
-	import ButtonHero from '$lib/components/ui/ButtonHero.svelte';
+	import ButtonHero from '$lib/components/hero/ButtonHero.svelte';
 	import { isBusy } from '$lib/derived/busy.derived';
 	import { i18n } from '$lib/stores/i18n.store';
 </script>
