+++ src/frontend/src/lib/components/receive/ReceiveButton.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import IconQr from '$lib/components/icons/IconQr.svelte';
-	import ButtonHero from '$lib/components/ui/ButtonHero.svelte';
+	import ButtonHero from '$lib/components/hero/ButtonHero.svelte';
 	import { RECEIVE_TOKENS_MODAL_OPEN_BUTTON } from '$lib/constants/test-ids.constants';
 	import { isBusy } from '$lib/derived/busy.derived';
 	import { i18n } from '$lib/stores/i18n.store';
